feat: align point cloud/3DGS with xodr + f16-pack splats to raise cap
Origin alignment:
- Add applyOrigin/applySplatOrigin in usePointCloudViewport so the
  origin-relative point cloud and 3DGS splat buffers are shifted back into
  the OpenDRIVE road's absolute frame (both Tauri and web paths), fixing the
  position offset between road_mesh.ply and the .xodr road network.

3DGS quality (f16 packing + raised caps):
- gaussian.rs: f32_to_f16_bits (RNE), build_splat_buffer_sh_f16 +
  sh_buffer_stride_f16. Position stays f32; cov/opacity/SH packed as f16
  pairs into u32 words (deg3: 232B -> 124B per splat, ~1.87x density).
- WASM gaussian_splat_buffer_sh returns Vec<u32>; native builds the packed
  buffer (u32s_to_bytes); meta.shStride uses the f16 stride.
- splatShader decodes array<u32> via unpack2x16float + bitcast<f32>.
- Raise PREVIEW_SPLAT_BUDGET 4M->16M and native load budget to 16M (web/WASM
  stays 4M for wasm32 heap safety); GPU storage cap now fits ~2x more splats.
- Frontend transport switched to Uint32Array end to end.

Tests: 4 Rust f16 tests, native gaussian tests, 1950 frontend tests pass;
typecheck clean.

+++ frontend/src/integration/rnkNextSdk.ts
@@ -228,11 +228,12 @@ export interface WorldEditorRenderer {
   loadGaussianSplats?(url: string): Promise<GaussianSplatInfo | undefined>;
   /**
    * Upload an already-parsed 3D Gaussian Splatting SH buffer (as produced by
-   * the Rust `gaussian_splat_buffer_sh`, `10 + (shDegree+1)²·3` floats/splat).
+   * the Rust `gaussian_splat_buffer_sh`, half-precision packed:
+   * `3 + ceil((7 + (shDegree+1)²·3)/2)` u32 words/splat).
    * Prefer {@link loadGaussianSplats} for the common fetch-and-render path.
    * Optional so hosts bundling an older SDK degrade gracefully.
    */
-  uploadGaussianSplats?(data: Float32Array, shDegree: number): void;
+  uploadGaussianSplats?(data: Uint32Array, shDegree: number): void;
   /** Remove the uploaded Gaussian splat cloud. */
   clearGaussianSplats?(): void;
 }
@@ -302,7 +303,7 @@ export interface WorldEditorWasm {
   /** Free a registered Gaussian splat cloud. */
   free_gaussian_splats(handle: number): void;
   /** View-dependent SH instance buffer (`10 + (shDegree+1)²·3` floats/splat). */
-  gaussian_splat_buffer_sh(handle: number): Float32Array;
+  gaussian_splat_buffer_sh(handle: number): Uint32Array;
   /** Summary `{ count, shDegree, shStride, origin, min, max }` of a splat cloud. */
   gaussian_splat_meta(handle: number): {
     count: number;
@@ -606,7 +607,7 @@ function adaptRenderer(wasm: WasmModule): WorldEditorRenderer {
         wasm.free_gaussian_splats(handle);
       }
     },
-    uploadGaussianSplats: (data: Float32Array, shDegree: number) => {
+    uploadGaussianSplats: (data: Uint32Array, shDegree: number) => {
       renderer.uploadGaussianSplats(data, shDegree);
     },
     clearGaussianSplats: () => {
